requirements.txt
# Core dependencies for Educational Antivirus Research Tool
# No external dependencies required for basic functionality
# All modules use Python standard library only

# Required dependencies:
pyyaml>=6.0  # For YAML configuration support

# Optional dependencies for enhanced functionality:
# click>=8.0   # For enhanced CLI interface
# rich>=13.0   # For better console output formatting